% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gendata}
\alias{gendata}
\title{gendata: generative mechanism for simulating synthetic data}
\usage{
gendata(
  n_obs = 200,
  p = 10,
  minb = 1.5,
  maxb = 3,
  error = 0.01,
  scenario = 4,
  nnnc = 3,
  noi = 3,
  ha = 2
)
}
\description{
gendata: generative mechanism for simulating synthetic data
}
