\name{invent-package}
\alias{invent-package}
\alias{invent}
\docType{package}
\title{
\packageTitle{invent}
}
\description{
\packageDescription{invent}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{invent}
\packageIndices{invent}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{invent}

Maintainer: \packageMaintainer{invent}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
