% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{invParMCMC}
\alias{invParMCMC}
\title{invParMCMC in Parallel}
\usage{
invParMCMC(
  y,
  x,
  hyperpar = c(3, 1, 1, 1, 0.00025, 0.4, 1.6, 0.2, 1.8, 0.4, 1.6, 0.2, 1.8),
  mht = c(1.4, 0.8, 1, 0.3, 0.7, 0.4, 4, 2.5),
  PredOutOfSample = TRUE,
  rank = 0.95,
  iter = 10000,
  burnin = iter/2,
  thin = 5,
  ha = 2,
  data = NULL,
  nchain = 2,
  percValidationSet = 20,
  seed = 10,
  onlyEff = TRUE
)
}
\description{
invParMCMC in Parallel
}
