% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{nlivcC}
\alias{nlivcC}
\title{nlicC}
\usage{
nlivcC(
  y,
  x,
  hyperpar = c(5, 25, 5, 5, 0.00025, 1, 1),
  mht = c(0.8, 1.4, 0.2, 1.4, 0.7, 0.15, 0.8, 0.4),
  rank = 0.95,
  iter = 10000,
  burnin = iter/2,
  thin = 5,
  ha
)
}
\description{
nlicC
}
