% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invent.R
\docType{package}
\name{nlvcm}
\alias{invent}
\alias{invent-package}
\alias{nlvcm}
\title{invent}
\description{
INteractions Varying coefficiENTs
}
\details{
The implementation has been done in \code{C++} through the use of
\code{Rcpp} and \code{RcppArmadillo}.
}
\references{
TBA
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidefabbrico/invent}
  \item Report bugs at \url{https://github.com/davidefabbrico/invent/issues}
}

}
\author{
Davide Fabbrico

Maintainer: Davide Fabbrico \email{davide.fabbrico@unifi.it}
}
